requirements.txt
customtkinter>=5.2.0
pyautogui>=0.9.54
keyboard>=0.13.5
pynput>=1.7.6
pillow>=10.0.0
pyinstaller>=5.13.0
